Fixed unsafe clone() of Response causing json() hang.
It's not safe to clone a stream unless the clones are processed in
parallel. However, the original response is not even being exposed
to consumers until the cloned stream is fully processed. Thus, we
can end up with a hang when handling large JSON responses,
particularly on NodeJS when using node-fetch. See:

https://github.com/node-fetch/node-fetch/issues/139
https://github.com/node-fetch/node-fetch/issues/665

This does have the side-effect that consumers will be unable to
rewind the stream and inspect the body in the event that JSON
parsing fails. If it succeeds, then we're provided with the parsed
body, so it's a non-issue in that case.